seg stupid me

+++ src/cellsociety/Segregation.java
@@ -12,8 +12,8 @@ public class Segregation extends simulation {
     public static final int TYPE_1 = 1;
     public static final int TYPE_2= 2;
     private int[][] myGrid;
-    public static int percent;
-    public static int totalN;
+    public static double percent;
+    public static double totalN;
     private int satis_Factor; //this will be given in the file.
 
 
@@ -37,7 +37,7 @@ public class Segregation extends simulation {
 
     }
 
-    public int getPercent(int i, int j, int type) {
+    public double getPercent(int i, int j, int type) {
             if (isBounds(i - 1, j) && myGrid[i - 1][j] != EMPTY) {
                 totalN++; //count the total neighbours
                 if (myGrid[i - 1][j] == type) percent++;
@@ -82,15 +82,11 @@ public class Segregation extends simulation {
 
 
     public void move(int type, int i, int j, int[][] futureState){
-        //FIND THE NEAREST EMPTY CELL AND MOVE
-
         for(int row=0; row<myGrid.length;row++){
             for(int col = 0; col< myGrid[0].length;col++){
                 if(myGrid[row][col] == EMPTY){
                     futureState[row][col] = type;
                     futureState[i][j] = EMPTY;
-
-
 
                 }
 
@@ -104,14 +100,16 @@ public class Segregation extends simulation {
         for (int i = 0; i < myGrid.length; i++) {
             for (int j = 0; j < myGrid[0].length; j++) {
                 if (myGrid[i][j] == TYPE_1) {
-                    int percentage = getPercent(i, j, TYPE_1);
+                    double percentage = getPercent(i, j, TYPE_1);
                     if (percentage < satis_Factor) {
                         move(TYPE_1, i ,j, futureState);
 
                     }
+
                 }
+
                 else if (myGrid[i][j] == TYPE_2) {
-                    int percentage = getPercent(i, j, TYPE_2);
+                    double percentage = getPercent(i, j, TYPE_2);
                     if (percentage < satis_Factor) {
                         move(TYPE_2, i ,j, futureState);
